Add tests for URL and Paginated, fix a float/int leak found by them
Both were real, uncovered logic: URL's HTTP_HOST/REQUEST_URI parsing
(the class underlying the Projects::searchProject() bug fixed
earlier) and Paginated's page-window/ellipsis math.

Paginated::paginate() built via reflection to skip its DB-touching
constructor (Model's defaults, no injectable PDO here) - same pattern
as this repo's own RenderTemplateTest-equivalent tests elsewhere, via
a new ConcretePaginated fixture (Paginated is abstract only because
of initAll()).

Found and fixed along the way: paginate()'s $lastPage came from
ceil(), which always returns float - so a page's own number could be
int (from the for loop) or float (the trailing/leading page from
$lastPage/$firstPage) depending on which branch pushed it, silently
inconsistent for anything comparing or serializing the result
strictly. Cast to int - ceil() of an int division is always a whole
number.

19 new tests, 187 total, all passing. PHPStan clean.

+++ src/Model/Paginated.php
@@ -87,7 +87,7 @@ abstract class Paginated extends Model
         $currentPage = $this->page + 1;
         $itemCount   = count($this->items);
         $firstPage   = 1;
-        $lastPage    = ceil($itemCount / $this->itemsPerPage);
+        $lastPage    = (int) ceil($itemCount / $this->itemsPerPage);
         if ($lastPage == 1 || $this->page > $lastPage) {
             // no pagination
             return array();
